requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
idna==3.3
python-decouple==3.6
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
